Get rid of pluggable menu bar
