+++ packages/jbrowse-desktop/src/sessionModelFactory.ts
@@ -43,9 +43,6 @@ export default function sessionModelFactory(pluginManager: any) {
           pluginManager.pluggableMstType('drawer widget', 'stateModel'),
         ),
       ),
-      menuBars: types.array(
-        pluginManager.pluggableMstType('menu bar', 'stateModel'),
-      ),
       connectionInstances: types.map(
         types.array(pluginManager.pluggableMstType('connection', 'stateModel')),
       ),
@@ -337,22 +334,6 @@ export default function sessionModelFactory(pluginManager: any) {
 
       hideAllDrawerWidgets() {
         self.activeDrawerWidgets.clear()
-      },
-
-      addMenuBar(
-        typeName: string,
-        initialState = {},
-        configuration = { type: typeName },
-      ) {
-        const typeDefinition = pluginManager.getElementType(
-          'menu bar',
-          typeName,
-        )
-        if (!typeDefinition)
-          throw new Error(`unknown menu bar type ${typeName}`)
-        const data = { ...initialState, type: typeName, configuration }
-        const model = typeDefinition.stateModel.create(data)
-        self.menuBars.push(model)
       },
 
       /**
