--- packages/menus/src/MainMenuBar/components/MainMenuBar.test.js
@@ -1,20 +1,0 @@
-import { createTestSession } from '@gmod/jbrowse-web/src/rootModel'
-import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles'
-import { render } from '@testing-library/react'
-import React from 'react'
-import MainMenuBar from './MainMenuBar'
-
-describe('<MainMenuBar />', () => {
-  it('renders', () => {
-    const session = createTestSession({
-      defaultSession: { menuBars: [{ id: 'testing', type: 'MainMenuBar' }] },
-    })
-    const model = session.menuBars[0]
-    const { container } = render(
-      <ThemeProvider theme={createMuiTheme()}>
-        <MainMenuBar model={model} />
-      </ThemeProvider>,
-    )
-    expect(container.firstChild).toMatchSnapshot()
-  })
-})
